Commit the base code for the new Universe Viewer
NOTE 1: Many thanks to the hard work done by the developers from Firestorm Viewer.  We have chosen to use the Firestorm Viewer code as the base for the new Universe Viewer based upon the needs of our community.

NOTE 2: We will be customizing this viewer and rebranding it very soon.  This will include dealing with any potential security issues found in the code.

Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>