Remove grids known to no longer be online
NOTE: This removes grids no longer online that we are aware of.  There may be grids still in the list that no longer are online or their information has changed that we are unaware of.  If for some reason you find one please open an issue ticket and let us know.  If you are a grid owner and would like your grid added to our grids list then please open a ticket and provide the necessary information for us to add your grid.
Signed-off-by: BriakallistaStarfinder <briakallista@secondgalaxy.com>