Bring our code base up to date with upstreams commits as of February 21, 2020.
NOTE: This will be the last time we do a major commit to get us caught up with upstream.  From now on all commits from Firestorm devs will be reviewed per our policies and procedures.

NOTE 2: Attention OpenSim users, there are known issues using any version later then Firestorm 6.0.2 for opensim due to the unwillingness of OpenSimulator's core devs to insure proper server side baking is in place to support Baked On Mesh.  While we respect their right to their decision, it is a very misguided decision and will only cause unnecessary use of computer resources in the viewer when it is something that should be done server-side.

Signed-off-by: BriakallistaStarfinder <briakallista@secondgalaxy.com>