Code rebase part 1 - Clean up the repository
- Remove the Firestorm 6.3.x codebase.  It had problems with OpenSim based grids anyway.

NOTE: We will be using the codebase from Angstrom Viewer which is a fork of Firestorm from Firestorm Version 4.3.2  It will be a compromise which allows us to implement the firestorm user interface that people appear to be more in favour of while allowing us to pick and choose what fixes, features, etc. we choose to implement from upstream.
Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>