Code Rebase Part 2 - Import the Angstrom Codebase
- Commit the Angstrom Viewer source code.

NOTE 1: Many thanks to Nhede for his hard work.  And many thanks to the Firestorm Viewer Development Team for their hard work.

NOTE 2: Angstrom Viewer is based off of Firestorm Viewer Version 4.3.2.  This was way back in January 2013 prior to Linden Labs removing the need for Curl from their viewers.

WARNING: Do not attempt to build this viewer yet.  It is not ready for building.  We first need to upgrade the build tools in order for things to be clean.
Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>