Rebrand Part 2 - Change file names
- Change the necessary file names to our viewer's name

WARNING: Do not attempt to build this viewer for the moment.  The build will fail.  By adhering to this advice you will help prevent reporting of issues with the build that we already know are occurring at the moment due to our need to upgrade the build tools which will happen in short order.
Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>