Rebrand Part 3 - Change viewer images and icons
- Change the necessary icons and images required by the viewer to ours

NOTE: If you want to contribute new and improved universe logos for the viewer please review this commit and make the necessary image sizes and file types.  Submit those images to Emperor Starfinder via discord or via a pull request here for review and legal approvals.

WARNING: Do not attempt to build this viewer for the moment.  The build will fail.  By adhering to this advice you will help prevent reporting of issues with the build that we already know are occurring at the moment due to our need to upgrade the build tools which will happen in short order.
Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>