Make sure the grids list only contains current grids by default
If you would like your grid added to our viewer you need to open an issue at: https://github.com/Virtual-Universe/Universe-Viewer-Early-Dev/issues and provide a copy of your grid's details.

WARNING: The viewer is not ready to build yet.  If you attempt to build the viewer currently you will get numerous errors.  We are aware of this and we will have the build tools upgraded in short order.
Signed-off-by: CaseyraeStarfinder <caseyrae@secondgalaxy.com>